Backend/requirements.txt
bcrypt==4.2.1
blinker==1.9.0
boto3==1.34.11
botocore==1.34.11
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
kubernetes==29.0.0
MarkupSafe==3.0.2
psycopg2==2.9.10
python-dotenv==1.0.0
pytz==2024.1
SQLAlchemy==2.0.37
typing_extensions==4.12.2
urllib3>=1.25.4,<1.27
Werkzeug==3.1.3
